refactor(test): inline codec vectors into their conformance suites
Each half of `vectors.ts` had exactly one consumer, and the provenance is
per-codec anyway: the two cite different test262 files. Only the `"foobar"`
byte array was shared, three lines now written once per suite.

Matches `Bech32m.test.ts`, which keeps its BIP-350 vectors inline.
